requirements.txt
fastapi>=0.116.0
pydantic>=2.10.6
sqlalchemy>=2.0.38
uvicorn>=0.34.0
scikit-learn>=1.6.1
pandas>=2.2.3
pyyaml>=6.0.2
tqdm>=4.67.1
numpy>=2.2.3
pytest>=8.3.4
setuptools>=75.8.0
